Updated to Pair 1.4 with 2 new modules
This project has been updated to Pair 1.4, which makes it possible to manage languages and international localization in a more adherent way to standards. All the modules related to the new features have been updated and a function has been introduced in the Tools module that allows you to update an existing project, the translation files and install the new modules with just one click. This module can be downloaded as ZIP package from an app built for Pair v1.4 (e.g. the new Pair example) and installed on a previous version project.